requirements.txt
aiogram>=3.15,<4
anthropic>=0.42,<1
PyGithub>=2.5,<3
gspread>=6.1,<7
google-auth>=2.36,<3
python-dotenv>=1.0,<2
